folder name and handle search
