+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Route } from 'react-router-dom'
-import { Home, Search } from './modules'
+import { Home, Search } from './components'
 import './App.css'
 
 export class App extends React.Component {
@@ -14,10 +14,11 @@ export class App extends React.Component {
         )} />
         <Route path='/search' render={({ history }) => (
           <Search
-            onSearchBook={(book) => {
-              this.searchBook(book)
-              history.push('/')
-            }} />
+            // onSearchBook={(book) => {
+            //   this.searchBook(book)
+            //   history.push('/')
+            // }} 
+            />
         )} />
       </div>
     )
