+++ src/components/index.js
@@ -1,0 +1,4 @@
+export * from './book.component';
+export * from './book-shelf.component';
+export * from './home.component';
+export * from './search.component';
