--- src/modules/index.js
@@ -1,4 +1,0 @@
-export * from './book';
-export * from './book-shelf';
-export * from './home';
-export * from './search';
